pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "website"
version = "0.1.0"
description = "Automation and build tools for my website"
authors = ["Rohail Taimour"]
license = "MIT"
packages = [
    { include = "ica_db_update" },
]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.0"
pyyaml = "^6.0.1"
jinja2 = "^3.1.2"
colorlog = "^6.7.0"
importlib-resources = "^6.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
invoke = "^1.6.0"
black = "^23.7.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
flake8 = "^6.1.0"


[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''
[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88